first 50 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	long long y[] = {1, 2};
+	int y[] = {1, 2};
 	int count;
 
 	printf("1, 2, ");
 	for (count = 2 ; count < 51 ; count++)
 	{
 		y[count] = y[count - 1] + y[count - 2];
-		printf("%lld", y[count]);
+		printf("%d", y[count]);
 		if (count < 49)
 			printf(", ");
 	}
